repo-sqale: container transformation + preparation for more ref tables

+++ repo/repo-sqale/src/main/java/com/evolveum/midpoint/repo/sqale/qmodel/common/MContainer.java
@@ -14,5 +14,5 @@ import java.util.UUID;
 public class MContainer {
 
     public UUID ownerOid;
-    public Integer cid;
+    public Long cid;
 }
